# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Tractive API
  version: 1.0.0
  description: API documentation for Tractive backend
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        activeRole:
          type: string
        isVerified:
          type: boolean
...
security:
  - bearerAuth: []
paths:
...
  /api/orders/{id}/status:
    patch:
      summary: Update order status or transport status
      description: Requires buyer/agent/transporter/admin with proper ownership. Uses activeRole.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, payment_pending, paid, delivered]
                transportStatus:
                  type: string
                  enum: [pending, picked, on_transit, delivered]
                transporter:
                  type: string
            example:
              status: "paid"
              transportStatus: "picked"
              transporter: "507f1f77bcf86cd799439015"
      responses:
        '200':
          description: Updated order
...
  /api/orders/{id}/transporters:
    get:
      summary: List available transporters for a paid, not-yet-shipped order
      description: Buyer (order owner), agent, or admin can fetch transporters only when payment has been approved and the order transport status is still `pending`. This endpoint currently returns transporter accounts that can be browsed for that order flow; it does not yet rank or pre-match transporters by route, size, or capacity.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transporters list for the order
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      orderId:
                        type: string
                      orderStatus:
                        type: string
                      paymentPendingApproval:
                        type: boolean
                      transportStatus:
                        type: string
                      assignedTransporter:
                        type: string
                        nullable: true
                      transporters:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
              example:
                success: true
                message: "Available transporters fetched"
                data:
                  orderId: "67deorder1234567890abc111"
                  orderStatus: "paid"
                  paymentPendingApproval: false
                  transportStatus: "pending"
                  assignedTransporter: null
                  transporters:
                    - _id: "67detrans1234567890abc001"
                      name: "North Haul Logistics"
                      email: "transporter@example.com"
                      phone: "+2348001112222"
                      businessName: "North Haul Logistics Ltd"
                      image: "https://example.com/transporter.jpg"
                      country: "Nigeria"
                      state: "Kaduna"
        '400':
          description: Payment is not approved yet or shipping has already started
        '403':
          description: Forbidden for non-owner buyer or unsupported role
...